pyproject.toml
[tool.poetry]
name = "autogcr"
version = "0.1.0"
description = ""
authors = ["codeblech <my.malikyash@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
nodriver = "^0.38.post1"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
hrequests = "^0.9.2"
google-genai = "^1.12.1"
python-docx = "^1.1.2"
markdown2 = "^2.5.3"
pillow = "^11.2.1"
beautifulsoup4 = "^4.13.4"
loguru = "^0.7.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
